refactor(swarm-node): relocate SwapWiring into the node crate
Move SwapWiring, its prepare/swap_event_sender/spawn surface, and the
swap-chequebook cashout helper out of the native-only builder and into
vertex-swarm-node::core, alongside PseudosettleWiring, so the embedded
launcher can reach it. spawn now takes a bare TaskExecutor and, on the
swap-chequebook path, the type-erased chain provider and address book as
plain values, rather than a builder-local InfrastructureContext and
SharedChainProvider.

The native builder imports SwapWiring from the node crate and threads the
chain provider through unchanged; the prepare -> with_swap_events -> spawn
sequence is behaviour-identical. Gate everything behind swap, with the
on-chain cashout behind the new swap-chequebook node feature.
